requirements.txt
flask
flask-sqlalchemy
python-whois
cryptography
email_validator
requests
python-owasp-zap-v2.4
ipaddress
datetime